scripts/training/requirements.txt
torch==2.1.0
transformers==4.37.0
peft==0.8.0
accelerate==0.27.2
datasets==2.16.1
bitsandbytes==0.43.1
sentencepiece==0.1.99
json-repair==0.28.4
PyYAML==6.0.1
tqdm==4.66.2
tensorboard==2.16.2
